% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piRFA.plot.R
\name{piRFA.plot}
\alias{piRFA.plot}
\title{PI-MIMIC Graphics: EPC and SEPC}
\usage{
piRFA.plot(
  resultados,
  cov,
  metric = "SEPC.ALL",
  theme_option = theme_minimal(),
  bar_colors = c(`Uniform DIF` = "blue", `Non-uniform DIF` = "red"),
  cutoff_line = 0.1
)
}
\arguments{
\item{resultados}{List of results generated by \code{\link{piRFA}}.}

\item{cov}{Name of the covariate used in \code{\link{piRFA}}.}

\item{metric}{Metric to plot: either `"EPC"` or `"SEPC.ALL"` (default is `"SEPC.ALL"`).}

\item{theme_option}{ggplot2 theme (default is `theme_minimal()`).}

\item{bar_colors}{Colors for the types of DIF (default blue and red, named `"Uniform DIF"` and `"Non-uniform DIF"`).}

\item{cutoff_line}{Horizontal line(s) indicating the DIF cutoff (default `0.10`). Can be a vector, e.g., `c(-0.10, 0.10)`.}
}
\value{
A bar plot of EPC or SEPC coefficients for uniform and non-uniform DIF.
}
\description{
The `piRFA.plot()` function generates bar plots of EPC (EPC and SEPC.ALL) coefficients for uniform and non-uniform DIF, facilitating the interpretation of DIF effects in items evaluated by the \code{\link{piRFA}} function.
}
\details{
\code{\link{piRFA.plot}} takes the output from \code{\link{piRFA}}, focused on the unstandardized (EPC) or standardized (SEPC.ALL) metric of the expected parameter change. The user must choose which metric to plot.
}
\examples{
set.seed(123)
Exmp1.data <- data.frame(
  grp = sample(0:1, 100, replace = TRUE),  # Group variable
  item1 = sample(1:5, 100, replace = TRUE),
  item2 = sample(1:5, 100, replace = TRUE),
  item3 = sample(1:5, 100, replace = TRUE))

Exmp1.output <- piRFA(data = Exmp1.data, items = c("Item1", "Item2", "Item3"), cov = "grp")

# Plot standardized expected parameter change (SEPC.ALL)
piRFA.plot(Exmp1.output, cov = "grp")

# Plot with a custom cutoff range
piRFA.plot(Exmp1.output, cov = "grp", cutoff_line = c(-.10, .10))

# Plot unstandardized expected parameter change (EPC)
piRFA.plot(Exmp1.output, cov = "grp", metric = "EPC")

}
\seealso{
\code{\link{ggplot2}}, \code{\link{piRFA}}
}
