% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piRFA.profile.R
\name{piRFA.profile}
\alias{piRFA.profile}
\title{Profile Plot of DIF for PI-MIMIC Output (Product of Indicators within a Multiple-Indicators Multiple-Causes Framework)}
\usage{
piRFA.profile(
  resultados,
  data,
  item,
  cov,
  theta_range = c(-3, 3),
  n_points = 200,
  theme_option = ggplot2::theme_minimal(),
  labels = NULL
)
}
\arguments{
\item{resultados}{Output object from the `piRFA()` function.}

\item{data}{The dataset used in the `piRFA()` analysis.}

\item{item}{The name of the item to be plotted (as a character string).}

\item{cov}{The name of the grouping covariate used in `piRFA()` (as a character string).}

\item{theta_range}{A numeric vector of length 2 specifying the latent trait range (default is -3 to 3).}

\item{n_points}{Number of evaluation points along the latent trait (default = 200).}

\item{theme_option}{A `ggplot2` theme object to customize the plot appearance.
Common options include: #' theme_minimal() (default), theme_bw(), theme_classic(),
theme_light(), theme_dark(), theme_void(), theme_gray(), theme_linedraw(), theme_test()
and others.}

\item{labels}{Optional character vector to rename groups in the plot legend.
If `NULL` (default), the original group names from `data[[cov]]` are used.}
}
\value{
A `ggplot2` object showing the DIF profile for PI-MIMIC output.
}
\description{
This function plots the expected response curves (linear prediction) across levels of the latent trait (θ)
for each group defined by a covariate. It is useful to visually inspect DIF—especially non-uniform DIF.
}
\details{
The function uses the parameter estimates from `piRFA()` to compute predicted responses
for each group along a latent variable continuum.

The formula used is a linear prediction:
\deqn{y = intercept + loading * theta + direct_effect * group + interaction * (theta × group)}
}
