% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piMIMIClrt.R
\name{piMIMIClrt}
\alias{piMIMIClrt}
\title{DIF analysis with PI-MIMIC using Likelihood Ratio Tests (LRT)}
\usage{
piMIMIClrt(data, items, cov, lvname = "LatFact", est = "MLM")
}
\arguments{
\item{data}{Data frame containing items and the covariate.}

\item{items}{Character vector of item names.}

\item{cov}{Name of the covariate (numeric or factor; factors are converted).}

\item{lvname}{Name of the latent variable (default "LatFact").}

\item{est}{Estimator for lavaan (default "MLM"; can be "ML", "ULS", etc.).}
}
\value{
A list with:
\item{DIF.Global}{Global test (2 df) for each item: Chi2, df, p.value.}
\item{DIF.Uniforme}{Uniform DIF test (1 df).}
\item{DIF.NoUniforme}{Non-uniform DIF test (1 df).}
\item{DeltaR2.uDIF}{Change in R² (unrestricted - restricted) for uniform DIF.}
\item{DeltaR2.nuDIF}{Change in R² for non-uniform DIF.}
\item{fit}{Unrestricted lavaan model object.}
}
\description{
Implements the product indicator (PI) approach for MIMIC models to detect
uniform and non-uniform DIF. Uses LRT between unrestricted and restricted
models (fixing DIF parameters to zero) and reports change in R² as effect size.
Follows the syntax and logic of Kolbe & Jorgensen (2018), Table 2.
}
\examples{
\dontrun{
library(psych)
data(bfi)
data.bfi <- bfi[, c("N1","N2","N3","N4","N5","gender")]
data.bfi <- data.bfi[complete.cases(data.bfi), ]
data.bfi$gender <- as.factor(data.bfi$gender)
neuro.items <- c("N1","N2","N3","N4","N5")
res <- piMIMIClrt(data = data.bfi, items = neuro.items,
                  cov = "gender", lvname = "Neuroticism", est = "MLM")
res$DIF.Global
res$DeltaR2.uDIF
}

}
