% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piRFA.plot.R
\name{piRFA.plot}
\alias{piRFA.plot}
\title{PI-MIMIC graphics: EPC and SEPC}
\usage{
piRFA.plot(
  resultados,
  cov,
  metric = "SEPC.ALL",
  theme_option = theme_minimal(),
  bar_colors = c(`DIF Uniforme` = "blue", `DIF No Uniforme` = "red"),
  punto_corte = 0.1
)
}
\arguments{
\item{resultados}{Lista de resultados generados por \code{\link{piRFA}}.}

\item{cov}{Nombre de la covariable usada en \code{\link{piRFA}}.}

\item{theme_option}{Tema para `ggplot2` (por defecto `theme_minimal()`).}

\item{bar_colors}{Colores para los tipos de DIF (por defecto azul y rojo).}

\item{punto_corte}{Línea vertical que indica el umbral de DIF (por defecto `0.10`). Puede incluirse un vector, e.g., `c(-0.10, 0.10)`.}
}
\value{
Un gráfico de barras con los coeficientes SEPC para DIF uniforme y no uniforme.
}
\description{
La función `piRFA.plot()` genera gráficos de los coeficientes EPC (EPC y SEPC.ALL) para DIF uniforme y no uniforme, 
facilitando la interpretación de los efectos DIF en los ítems evaluados por la función \code{\link{piRFA}}.
}
\details{
\code{\link{piRFA.plot}} toma la salida de \code{\link{piRFA}}, centrada en el unstandardized (epc) o standardized
metric of the expected parameter change (spec.all). El usuario debe elegir cual quiere graficar.
}
\examples{
### Example 1 -------------

# Example 1 data
set.seed(123)
Exmp1.data <- data.frame(
 grp = sample(0:1, 100, replace = TRUE),  # Variable de grupo
 item1 = sample(1:5, 100, replace = TRUE),
 item2 = sample(1:5, 100, replace = TRUE),
 item3 = sample(1:5, 100, replace = TRUE))

# Output and save
Exmp1.output <- piRFA(data = Exmp1.data , items = c("Item1", "Item2", "Item3"), cov = "grp")

# Create graph
piRFA.plot(Exmp1.output, cov = "grp")

# Add cutoff range
piRFA.plot(Exmp1.output, cov = "grp", punto_corte = c(-.10, .10))


}
\seealso{
\code{\link{ggplot2}}, \code{\link{piRFA}}
}
