% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piRFA.R
\name{piRFA}
\alias{piRFA}
\title{Function to Analyze DIF with PI-MIMIC (Product of Indicators)}
\usage{
piRFA(data, items, cov, lvname = "LatFact", est = "MLM")
}
\arguments{
\item{data}{DataFrame containing items and the covariate.}

\item{items}{Vector of item names within `data`.}

\item{cov}{Name of the covariate in `data` (can be categorical or numeric). If categorical, it must be a factor.}

\item{lvname}{Name for the latent variable in the model (default is `"LatFact"`).}

\item{est}{Abbreviation of the estimator to use; options are: "MLM", "MLR", "ULSMV" (default is `"MLM"`).}
}
\value{
The function returns a list with the following DataFrames:
 \itemize{
   \item \code{DIF_Global} - Global DIF results.
   \item \code{DIF_Uniforme} - Uniform DIF results.
   \item \code{DIF_NoUniforme} - Non-uniform DIF results.
   \item \code{SEPC} - SEPC coefficients for uniform and non-uniform DIF.
 }
}
\description{
The `piRFA` function analyzes Differential Item Functioning (DIF)
using the multiple-indicators multiple-causes (MIMIC) framework with product of indicators (PI).
It relies on `lavaan` and `scripty`. Uniform and non-uniform DIF can be evaluated in a measurement scale,
through statistical tests and an effect size approximation.
}
\details{
This function implements Differential Item Functioning (DIF) analysis using
the Product of Indicators approach (PI; Kolbe & Jorgensen, 2018) within the
Restricted Factor Analysis (RFA; Oort, 1998) framework. This method operates under a
MIMIC scheme (Finch, 2005), incorporating latent variable interactions using PI (Kolbe et al., 2018, 2019;
Kolbe, Jorgensen, & Molenaar, 2020). It allows for the evaluation of uniform (uDIF) and non-uniform DIF (nuDIF)
with covariates that can be categorical (e.g., sex) or continuous (e.g., self-esteem, conscientiousness).

Estimation is performed via `lavaan::cfa`, and DIF statistical tests are based
on the likelihood ratio test (LRT), comparing an unrestricted model
(with covariate effect and interaction parameters freely estimated) and a restricted model
(with these parameters fixed to zero), using a chi-square test with 2 degrees of freedom (`df = 2`).

The implementation relies on `lavaan::lavTestScore` and `lavaan::cfa`, integrating modified functionalities
from `scripty::mimic` and `scripty::prods`. Unlike `scripty`, this function
avoids creating external files, allowing all output to be accessed in the R session.

As an effect size indicator, `piRFA()` uses the releasing fixed-to-zero
parameters approach (Whittaker, 2012; Garnier-Villarreal & Jorgensen, 2024). The expected parameter change
(`epc`) and the standardized expected parameter change (`sepc.all`) are estimated following Chou & Bentler (1993).
`sepc.all` represents the standardized unit change upon releasing the zero restriction
in the analyzed item parameters. Values `sepc.all ≥ .20` are considered indicative of
possible model misspecification (Whittaker, 2012), while values `≥ .10`
(Kaplan, 1989) are a threshold for a moderate effect.

The user can select the appropriate estimator among `"MLM"`, `"MLR"` and `"ULSMV"`,
depending on normality and robustness assumptions.
}
\examples{
### Example 1 -------------

set.seed(123)
Exmp1.data <- data.frame(
  grp = sample(0:1, 100, replace = TRUE),  # Group variable
  item1 = sample(1:5, 100, replace = TRUE),
  item2 = sample(1:5, 100, replace = TRUE),
  item3 = sample(1:5, 100, replace = TRUE),
  item4 = sample(1:5, 100, replace = TRUE))

# Run DIF analysis, and full output
piRFA(data = Exmp1.data , items = c("item1", "item2", "item3"), cov = "grp")

# Specific output: Uniform DIF (uDIF)
Exmp1.output <- piRFA(data = Exmp1.data , items = c("item1", "item2", "item3"), cov = "grp")

Exmp1.output$DIF_Uniforme

# Specific output: full SEPC
piRFA(data = Exmp1.data , items = c("item1", "item2", "item3"), cov = "grp")$SEPC

# Specific output: full SEPC for uniform DIF
piRFA(data = Exmp1.data , items = c("item1", "item2", "item3"), cov = "grp")$SEPC[c(1,3), ]



}
\references{
Chou, C.-P., & Bentler, P. (1993). Invariant standardized estimated parameter change for model modification in covariance structure analysis. *Multivariate Behavioral Research*, 28, 97–110. https://doi.org/10.1207/s15327906mbr2801_6

Finch, H. (2005). The MIMIC model as a method for detecting DIF: Comparison with Mantel–Haenszel, SIBTEST, and the IRT likelihood ratio. *Applied Psychological Measurement*, 29, 278–295. https://doi.org/10.1177/0146621605275728

Garnier-Villarreal, M., & Jorgensen, T. D. (2024). Evaluating Local Model Misspecification with Modification Indices in Bayesian Structural Equation Modeling. *Structural Equation Modeling: A Multidisciplinary Journal*, 1–15. https://doi.org/10.1080/10705511.2024.2413128

Kaplan D. (1989). Model Modification in Covariance Structure Analysis: Application of the Expected Parameter Change Statistic. *Multivariate Behavioral Research*, 24(3), 285–305. https://doi.org/10.1207/s15327906mbr2403_2

Kolbe, L., & Jorgensen, T. D. (2018). Using product indicators in restricted factor analysis models to detect nonuniform measurement bias. In M. Wiberg, S. A. Culpepper, R. Janssen, J. González, & D. Molenaar (Eds.), Quantitative psychology: The 82nd Annual Meeting of the Psychometric Society, Zurich, Switzerland, 2017 (pp. 235–245). New York, NY: Springer. https://doi.org/10.1007/978-3-319-77249-3_20

Kolbe, L., & Jorgensen, T. D. (2019). Using restricted factor analysis to select anchor items and detect differential item functioning. *Behavior Research Methods*, 51, 138–151. https://doi.org/10.3758/s13428-018-1151-3

Kolbe, L., Jorgensen, T. D., & Molenaar, D. (2020). The Impact of Unmodeled Heteroskedasticity on Assessing Measurement Invariance in Single-group Models. *Structural Equation Modeling: A Multidisciplinary Journal*, 28(1), 82–98. https://doi.org/10.1080/10705511.2020.1766357

Whittaker, T. A. (2012). Estimation of Standardized Expected Parameter Change for DIF Detection. *Educational and Psychological Measurement*, 72(3), 342-357.
}
\seealso{
\code{\link[lavaan]{cfa}}, \code{\link[scripty]{mimic}}, \code{\link[scripty]{prods}}, \code{\link{piRFA.plot}}
}
